Random Ray Testing Simplification (#3061)
