--- tests/regression_tests/random_ray_cube/False_hybrid/results_true.dat
@@ -1,9 +1,0 @@
-tally 1:
--5.920549E+04
-2.321317E+09
-tally 2:
-4.761087E+02
-4.909570E+04
-tally 3:
-2.206028E+01
-1.019164E+02
