--- tests/regression_tests/random_ray_cube/True_hybrid/results_true.dat
@@ -1,9 +1,0 @@
-tally 1:
--5.230164E+02
-1.818988E+05
-tally 2:
-3.067508E-02
-2.037701E-04
-tally 3:
-1.941220E-03
-7.892297E-07
