--- tests/regression_tests/random_ray_fixed_source/cell/inputs_true.dat
@@ -1,205 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<model>
-  <materials>
-    <cross_sections>mgxs.h5</cross_sections>
-    <material id="1" name="source">
-      <density units="macro" value="1.0"/>
-      <macroscopic name="source"/>
-    </material>
-    <material id="2" name="void">
-      <density units="macro" value="1.0"/>
-      <macroscopic name="void"/>
-    </material>
-    <material id="3" name="shield">
-      <density units="macro" value="1.0"/>
-      <macroscopic name="shield"/>
-    </material>
-  </materials>
-  <geometry>
-    <cell id="1" material="1" name="infinite source region" universe="1"/>
-    <cell id="2" material="2" name="infinite void region" universe="2"/>
-    <cell id="3" material="3" name="infinite shield region" universe="3"/>
-    <cell fill="4" id="4" name="source lattice cell" universe="7"/>
-    <cell fill="5" id="5" name="void lattice cell" universe="8"/>
-    <cell fill="6" id="6" name="shield lattice cell" universe="9"/>
-    <cell fill="10" id="7" name="dogleg lattice cell" universe="11"/>
-    <cell fill="11" id="8" name="full domain" region="1 -2 3 -4 5 -6" universe="12"/>
-    <lattice id="4">
-      <pitch>5.0 5.0 5.0</pitch>
-      <dimension>2 2 2</dimension>
-      <lower_left>-5.0 -5.0 -5.0</lower_left>
-      <universes>
-1 1 
-1 1 
-
-1 1 
-1 1 </universes>
-    </lattice>
-    <lattice id="5">
-      <pitch>5.0 5.0 5.0</pitch>
-      <dimension>2 2 2</dimension>
-      <lower_left>-5.0 -5.0 -5.0</lower_left>
-      <universes>
-2 2 
-2 2 
-
-2 2 
-2 2 </universes>
-    </lattice>
-    <lattice id="6">
-      <pitch>5.0 5.0 5.0</pitch>
-      <dimension>2 2 2</dimension>
-      <lower_left>-5.0 -5.0 -5.0</lower_left>
-      <universes>
-3 3 
-3 3 
-
-3 3 
-3 3 </universes>
-    </lattice>
-    <lattice id="10">
-      <pitch>10.0 10.0 10.0</pitch>
-      <dimension>6 10 6</dimension>
-      <lower_left>0.0 0.0 0.0</lower_left>
-      <universes>
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-8 8 8 8 9 9 
-8 9 9 9 9 9 
-8 9 9 9 9 9 
-8 9 9 9 9 9 
-8 9 9 9 9 9 
-7 9 9 9 9 9 
-
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 8 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 8 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-
-9 9 9 8 9 9 
-9 9 9 8 9 9 
-9 9 9 8 9 9 
-9 9 9 8 9 9 
-9 9 9 8 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 
-9 9 9 9 9 9 </universes>
-    </lattice>
-    <surface boundary="reflective" coeffs="0.0" id="1" type="x-plane"/>
-    <surface boundary="vacuum" coeffs="60.0" id="2" type="x-plane"/>
-    <surface boundary="reflective" coeffs="0.0" id="3" type="y-plane"/>
-    <surface boundary="vacuum" coeffs="100.0" id="4" type="y-plane"/>
-    <surface boundary="reflective" coeffs="0.0" id="5" type="z-plane"/>
-    <surface boundary="vacuum" coeffs="60.0" id="6" type="z-plane"/>
-  </geometry>
-  <settings>
-    <run_mode>fixed source</run_mode>
-    <particles>1000</particles>
-    <batches>10</batches>
-    <inactive>5</inactive>
-    <source particle="neutron" strength="1.0" type="independent">
-      <energy type="discrete">
-        <parameters>100.0 1.0</parameters>
-      </energy>
-      <constraints>
-        <domain_type>cell</domain_type>
-        <domain_ids>4</domain_ids>
-      </constraints>
-    </source>
-    <energy_mode>multi-group</energy_mode>
-    <random_ray>
-      <distance_active>400.0</distance_active>
-      <distance_inactive>100.0</distance_inactive>
-      <volume_normalized_flux_tallies>False</volume_normalized_flux_tallies>
-      <source particle="neutron" strength="1.0" type="independent">
-        <space type="box">
-          <parameters>0.0 0.0 0.0 60.0 100.0 60.0</parameters>
-        </space>
-      </source>
-    </random_ray>
-  </settings>
-  <tallies>
-    <mesh id="1">
-      <dimension>1 10 1</dimension>
-      <lower_left>0.0 0.0 0.0</lower_left>
-      <upper_right>10.0 100.0 10.0</upper_right>
-    </mesh>
-    <mesh id="2">
-      <dimension>6 1 1</dimension>
-      <lower_left>0.0 50.0 0.0</lower_left>
-      <upper_right>60.0 60.0 10.0</upper_right>
-    </mesh>
-    <mesh id="3">
-      <dimension>6 1 1</dimension>
-      <lower_left>0.0 90.0 30.0</lower_left>
-      <upper_right>60.0 100.0 40.0</upper_right>
-    </mesh>
-    <filter id="1" type="mesh">
-      <bins>1</bins>
-    </filter>
-    <filter id="2" type="mesh">
-      <bins>2</bins>
-    </filter>
-    <filter id="3" type="mesh">
-      <bins>3</bins>
-    </filter>
-    <tally id="1" name="Case 3A">
-      <filters>1</filters>
-      <scores>flux</scores>
-      <estimator>analog</estimator>
-    </tally>
-    <tally id="2" name="Case 3B">
-      <filters>2</filters>
-      <scores>flux</scores>
-      <estimator>analog</estimator>
-    </tally>
-    <tally id="3" name="Case 3C">
-      <filters>3</filters>
-      <scores>flux</scores>
-      <estimator>analog</estimator>
-    </tally>
-  </tallies>
-</model>
