--- tests/regression_tests/random_ray_fixed_source/material/results_true.dat
@@ -1,47 +1,0 @@
-tally 1:
-3.751047E+01
-2.841797E+02
-1.000671E+01
-2.020214E+01
-3.979569E+00
-3.330838E+00
-2.552971E+00
-1.407542E+00
-1.736764E+00
-7.948004E-01
-1.178834E+00
-4.328006E-01
-6.953895E-01
-1.408391E-01
-2.389347E-01
-1.617450E-02
-8.128516E-02
-1.903860E-03
-2.764062E-02
-2.285458E-04
-tally 2:
-1.178834E+00
-4.328006E-01
-3.982124E-01
-4.138078E-02
-9.676374E-02
-1.919297E-03
-3.944531E-02
-3.133733E-04
-1.554518E-02
-5.084822E-05
-4.815079E-03
-4.681450E-06
-tally 3:
-1.635823E-03
-6.482611E-07
-1.248235E-03
-3.227247E-07
-1.248195E-03
-3.488233E-07
-1.030935E-03
-2.206418E-07
-5.670315E-04
-6.648507E-08
-2.393638E-04
-1.242517E-08
