+++ tests/regression_tests/random_ray_fixed_source_domain/universe/results_true.dat
@@ -1,0 +1,9 @@
+tally 1:
+-5.745886E+02
+9.758367E+04
+tally 2:
+2.971927E-02
+1.827222E-04
+tally 3:
+1.978393E-03
+7.951531E-07
