+++ tests/regression_tests/random_ray_fixed_source_normalization/False/results_true.dat
@@ -1,0 +1,9 @@
+tally 1:
+-6.614590E+04
+1.283943E+09
+tally 2:
+4.612657E+02
+4.402080E+04
+tally 3:
+2.248302E+01
+1.026884E+02
