+++ tests/regression_tests/random_ray_k_eff/results_true.dat
