+++ tests/regression_tests/random_ray_k_eff/test.py
@@ -1,0 +1,19 @@
+import os
+
+from openmc.examples import random_ray_lattice
+
+from tests.testing_harness import TolerantPyAPITestHarness
+
+
+class MGXSTestHarness(TolerantPyAPITestHarness):
+    def _cleanup(self):
+        super()._cleanup()
+        f = 'mgxs.h5'
+        if os.path.exists(f):
+            os.remove(f)
+
+
+def test_random_ray_basic():
+    model = random_ray_lattice()
+    harness = MGXSTestHarness('statepoint.10.h5', model)
+    harness.main()
