--- tests/regression_tests/random_ray_vacuum/results_true.dat
@@ -1,171 +1,0 @@
-k-combined:
-1.010455E-01 1.585558E-02
-tally 1:
-7.466984E-01
-1.245920E-01
-2.771079E-01
-1.714504E-02
-6.744252E-01
-1.015566E-01
-1.634870E-01
-6.288701E-03
-2.405120E-02
-1.362874E-04
-5.853580E-02
-8.072823E-04
-1.641162E-02
-6.568765E-05
-5.223801E-04
-6.753516E-08
-1.271369E-03
-4.000365E-07
-3.014736E-02
-1.922973E-04
-9.765325E-04
-2.043645E-07
-2.376685E-03
-1.210529E-06
-1.562379E-01
-4.906526E-03
-1.746664E-03
-6.141658E-07
-4.251084E-03
-3.638028E-06
-1.826385E+00
-6.678141E-01
-3.095716E-03
-1.920117E-06
-7.660132E-03
-1.175650E-05
-2.013809E+00
-8.136726E-01
-3.015545E-02
-1.832201E-04
-8.387586E-02
-1.417475E-03
-1.573069E+00
-5.388374E-01
-0.000000E+00
-0.000000E+00
-0.000000E+00
-0.000000E+00
-2.867849E-01
-1.864798E-02
-0.000000E+00
-0.000000E+00
-0.000000E+00
-0.000000E+00
-2.136372E-02
-1.035499E-04
-0.000000E+00
-0.000000E+00
-0.000000E+00
-0.000000E+00
-3.973323E-02
-3.229766E-04
-0.000000E+00
-0.000000E+00
-0.000000E+00
-0.000000E+00
-1.779974E-01
-6.350613E-03
-0.000000E+00
-0.000000E+00
-0.000000E+00
-0.000000E+00
-1.036886E+00
-2.151147E-01
-0.000000E+00
-0.000000E+00
-0.000000E+00
-0.000000E+00
-1.109991E+00
-2.468005E-01
-0.000000E+00
-0.000000E+00
-0.000000E+00
-0.000000E+00
-5.813217E-01
-7.704274E-02
-2.160141E-01
-1.062660E-02
-5.257350E-01
-6.294540E-02
-1.358032E-01
-4.462239E-03
-2.030839E-02
-1.002755E-04
-4.942655E-02
-5.939703E-04
-1.612888E-02
-6.604109E-05
-5.274425E-04
-7.250333E-08
-1.283689E-03
-4.294649E-07
-2.985723E-02
-1.925418E-04
-9.868482E-04
-2.142916E-07
-2.401791E-03
-1.269331E-06
-1.590678E-01
-5.110940E-03
-1.800903E-03
-6.566490E-07
-4.383091E-03
-3.889678E-06
-1.928611E+00
-7.475889E-01
-3.310101E-03
-2.209547E-06
-8.190613E-03
-1.352862E-05
-2.172777E+00
-9.544513E-01
-3.334566E-02
-2.267149E-04
-9.274926E-02
-1.753971E-03
-7.566911E-01
-1.278105E-01
-2.881892E-01
-1.854375E-02
-7.013949E-01
-1.098417E-01
-1.662732E-01
-6.495387E-03
-2.515274E-02
-1.488430E-04
-6.121675E-02
-8.816538E-04
-1.682747E-02
-6.933017E-05
-5.514441E-04
-7.567901E-08
-1.342104E-03
-4.482757E-07
-3.068773E-02
-1.997062E-04
-1.021094E-03
-2.240938E-07
-2.485139E-03
-1.327393E-06
-1.578722E-01
-5.013656E-03
-1.812622E-03
-6.620083E-07
-4.411613E-03
-3.921424E-06
-1.922798E+00
-7.400474E-01
-3.401690E-03
-2.320513E-06
-8.417243E-03
-1.420805E-05
-2.178318E+00
-9.546106E-01
-3.451316E-02
-2.421994E-04
-9.599661E-02
-1.873766E-03
